requirements.txt
sqlalchemy
python-dotenv
psycopg2-binary
flask
requests
fastapi
uvicorn
APScheduler
